fix: resolucion de problema de dependencias
Detalles:
- Se ha resuelto problemasde dependencias que hubo durante los  ultimos
  merge
- S ha renombrado algunos directorios para mas consistencia
