+++ src/main/java/com/dk/dermokometicapi/mappers/ArticleMapper.java
@@ -3,9 +3,9 @@ package com.dk.dermokometicapi.mappers;
 import com.dk.dermokometicapi.models.dto.ArticleRequestDTO;
 import com.dk.dermokometicapi.models.dto.ArticleResponseDTO;
 import com.dk.dermokometicapi.models.dto.ArticleSummaryResponseDTO;
-import com.dk.dermokometicapi.models.entity.Article;
-import com.dk.dermokometicapi.models.entity.ArticleDetail;
-import com.dk.dermokometicapi.models.entity.Writer;
+import com.dk.dermokometicapi.models.entities.Article;
+import com.dk.dermokometicapi.models.entities.ArticleDetail;
+import com.dk.dermokometicapi.models.entities.Writer;
 import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Component;
