+++ src/main/java/com/dk/dermokometicapi/models/entities/User.java
@@ -1,4 +1,4 @@
-package com.dk.dermokometicapi.models.entity;
+package com.dk.dermokometicapi.models.entities;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,21 +8,22 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name = "writers")
-public class Writer {
+@Table(name = "users")
+public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 50)
-    private String name;
+    @Column(unique = true, nullable = false, length = 20)
+    private String username;
+
+    @Column(nullable = false, length = 20)
+    private String password;
 
     @Column(nullable = false, length = 50)
-    private String lastName;
-
-    @Column(nullable = true, length = 500)
-    private String description;
+    private String email;
 
     @Column(columnDefinition = "TEXT")
     private String profilePic;
+
 }
