+++ src/main/java/com/dk/dermokometicapi/models/entities/Writer.java
@@ -1,4 +1,4 @@
-package com.dk.dermokometicapi.models.entity;
+package com.dk.dermokometicapi.models.entities;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,22 +8,21 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-@Table(name = "users")
-public class User {
+@Table(name = "writers")
+public class Writer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(unique = true, nullable = false, length = 20)
-    private String username;
-
-    @Column(nullable = false, length = 20)
-    private String password;
+    @Column(nullable = false, length = 50)
+    private String name;
 
     @Column(nullable = false, length = 50)
-    private String email;
+    private String lastName;
+
+    @Column(nullable = true, length = 500)
+    private String description;
 
     @Column(columnDefinition = "TEXT")
     private String profilePic;
-
 }
