+++ src/main/java/com/dk/dermokometicapi/repositories/AnswerLikeRepository.java
@@ -1,8 +1,8 @@
 package com.dk.dermokometicapi.repositories;
 
-import com.dk.dermokometicapi.models.entity.Answer;
-import com.dk.dermokometicapi.models.entity.AnswerLike;
-import com.dk.dermokometicapi.models.entity.User;
+import com.dk.dermokometicapi.models.entities.Answer;
+import com.dk.dermokometicapi.models.entities.AnswerLike;
+import com.dk.dermokometicapi.models.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AnswerLikeRepository extends JpaRepository<AnswerLike, Long> {
