+++ src/main/java/com/dk/dermokometicapi/repositories/ArticleDetailRepository.java
@@ -1,6 +1,6 @@
 package com.dk.dermokometicapi.repositories;
 
-import com.dk.dermokometicapi.models.entity.ArticleDetail;
+import com.dk.dermokometicapi.models.entities.ArticleDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArticleDetailRepository extends JpaRepository<ArticleDetail, Long>{
