+++ src/main/java/com/dk/dermokometicapi/repositories/QuestionLikeRepository.java
@@ -1,8 +1,8 @@
 package com.dk.dermokometicapi.repositories;
 
-import com.dk.dermokometicapi.models.entity.Question;
-import com.dk.dermokometicapi.models.entity.QuestionLike;
-import com.dk.dermokometicapi.models.entity.User;
+import com.dk.dermokometicapi.models.entities.Question;
+import com.dk.dermokometicapi.models.entities.QuestionLike;
+import com.dk.dermokometicapi.models.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface QuestionLikeRepository extends JpaRepository<QuestionLike, Long> {
