+++ src/main/java/com/dk/dermokometicapi/services/AnswerService.java
@@ -3,10 +3,10 @@ package com.dk.dermokometicapi.services;
 import com.dk.dermokometicapi.exceptions.BadRequestException;
 import com.dk.dermokometicapi.exceptions.ResourceNotFoundException;
 import com.dk.dermokometicapi.models.dto.*;
-import com.dk.dermokometicapi.models.entity.Answer;
-import com.dk.dermokometicapi.models.entity.AnswerLike;
-import com.dk.dermokometicapi.models.entity.Question;
-import com.dk.dermokometicapi.models.entity.User;
+import com.dk.dermokometicapi.models.entities.Answer;
+import com.dk.dermokometicapi.models.entities.AnswerLike;
+import com.dk.dermokometicapi.models.entities.Question;
+import com.dk.dermokometicapi.models.entities.User;
 import com.dk.dermokometicapi.mappers.AnswerLikeMapper;
 import com.dk.dermokometicapi.mappers.AnswerMapper;
 import com.dk.dermokometicapi.repositories.AnswerLikeRepository;
