+++ src/main/java/com/dk/dermokometicapi/services/UserService.java
@@ -6,7 +6,7 @@ import com.dk.dermokometicapi.models.dto.UserRequestDTO;
 import com.dk.dermokometicapi.models.dto.UserResponseDTO;
 import com.dk.dermokometicapi.models.dto.UserUpdateDTO;
 import com.dk.dermokometicapi.models.dto.UserValidationDTO;
-import com.dk.dermokometicapi.models.entity.User;
+import com.dk.dermokometicapi.models.entities.User;
 import com.dk.dermokometicapi.mappers.UserMapper;
 import com.dk.dermokometicapi.repositories.UserRepository;
 import lombok.AllArgsConstructor;
