+++ src/test/java/com/dk/dermokometicapi/services/AnswerServiceTest.java
@@ -7,10 +7,9 @@ import com.dk.dermokometicapi.exceptions.ResourceNotFoundException;
 import com.dk.dermokometicapi.mappers.AnswerMapper;
 import com.dk.dermokometicapi.models.dto.AnswerRequestDTO;
 import com.dk.dermokometicapi.models.dto.AnswerResponseDTO;
-import com.dk.dermokometicapi.models.dto.ListRequestDTO;
-import com.dk.dermokometicapi.models.entity.Answer;
-import com.dk.dermokometicapi.models.entity.Question;
-import com.dk.dermokometicapi.models.entity.User;
+import com.dk.dermokometicapi.models.entities.Answer;
+import com.dk.dermokometicapi.models.entities.Question;
+import com.dk.dermokometicapi.models.entities.User;
 import com.dk.dermokometicapi.repositories.AnswerLikeRepository;
 import com.dk.dermokometicapi.repositories.AnswerRepository;
 import com.dk.dermokometicapi.repositories.QuestionRepository;
@@ -20,13 +19,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
 
-import java.awt.print.Pageable;
-import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
