+++ src/test/java/com/dk/dermokometicapi/services/QuestionServiceTest.java
@@ -1,22 +1,11 @@
 package com.dk.dermokometicapi.services;
 
-import com.dk.dermokometicapi.model.dto.QuestionLikeRequestDTO;
-import com.dk.dermokometicapi.model.dto.QuestionLikeResponseDTO;
-import com.dk.dermokometicapi.model.dto.QuestionRequestDTO;
-import com.dk.dermokometicapi.model.dto.QuestionResponseDTO;
-import com.dk.dermokometicapi.model.entity.Question;
-import com.dk.dermokometicapi.model.entity.QuestionLike;
-import com.dk.dermokometicapi.model.entity.User;
-import com.dk.dermokometicapi.model.exception.BadRequestException;
-import com.dk.dermokometicapi.model.exception.ResourceNotFoundException;
-import com.dk.dermokometicapi.model.mapper.QuestionLikeMapper;
-import com.dk.dermokometicapi.model.mapper.QuestionMapper;
-import com.dk.dermokometicapi.model.repository.QuestionLikeRepository;
-import com.dk.dermokometicapi.model.repository.QuestionRepository;
-import com.dk.dermokometicapi.model.repository.UserRepository;
-import com.dk.dermokometicapi.model.service.ArticleService;
-import com.dk.dermokometicapi.model.service.QuestionService;
-import com.dk.dermokometicapi.model.service.UserService;
+import com.dk.dermokometicapi.models.dto.*;
+import com.dk.dermokometicapi.models.entities.*;
+import com.dk.dermokometicapi.exceptions.*;
+import com.dk.dermokometicapi.mappers.*;
+import com.dk.dermokometicapi.repositories.*;
+import com.dk.dermokometicapi.services.*;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
