+++ src/test/java/com/dk/dermokometicapi/services/WriterServiceTest.java
@@ -1,6 +1,6 @@
 package com.dk.dermokometicapi.services;
 
-import com.dk.dermokometicapi.models.entity.Writer;
+import com.dk.dermokometicapi.models.entities.Writer;
 import com.dk.dermokometicapi.mappers.WriterMapper;
 import com.dk.dermokometicapi.repositories.WriterRepository;
 import org.junit.jupiter.api.Test;
